requirements.txt
scapy
PySide6
pandas
reportlab
matplotlib
pyyaml
pyshark
